ch 4 Node Modules - create module
