+++ matrices/courses/linkedin-learning/js/node/node-essential-training/practice/4-node-modules/lib/collectAnswers.js
@@ -1,0 +1,22 @@
+const readline = require('readline');
+
+const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout
+});
+
+module.exports = (questions, done = f => f) => {
+    const answers = [];
+    const [firstQuestion] = questions;
+
+    const questionAnswered = answer => {
+        answers.push(answer);
+        if (answers.length < questions.length) {
+            rl.question(questions[answers.length], questionAnswered);
+        } else {
+            done(answers);
+        }
+    };
+
+    rl.question(firstQuestion, questionAnswered);
+};
